Move attribute resolving theme related helpers to Theme.kt
