+++ anko/library/static/commons/src/Dimensions.kt
@@ -19,11 +19,7 @@ package org.jetbrains.anko
 
 import android.app.Fragment
 import android.content.Context
-import android.os.Build
-import android.support.annotation.AttrRes
-import android.support.annotation.ColorInt
 import android.support.annotation.DimenRes
-import android.util.TypedValue
 import android.view.View
 
 const val LDPI: Int = android.util.DisplayMetrics.DENSITY_LOW
@@ -52,26 +48,6 @@ fun Context.px2sp(px: Int): Float = px.toFloat() / resources.displayMetrics.scal
 
 fun Context.dimen(@DimenRes resource: Int): Int = resources.getDimensionPixelSize(resource)
 
-fun Context.attr(@AttrRes attribute: Int): TypedValue {
-    var typed = TypedValue()
-    ctx.theme.resolveAttribute(attribute, typed, true)
-    return typed
-}
-
-//returns px
-fun Context.dimenAttr(@AttrRes attribute: Int): Int = TypedValue.complexToDimensionPixelSize(attr(attribute).data, resources.displayMetrics)
-
-//returns color
-@ColorInt
-fun Context.colorAttr(@AttrRes attribute: Int): Int {
-    return if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.M){
-        resources.getColor(attr(attribute).resourceId, ctx.theme)
-    } else {
-        resources.getColor(attr(attribute).resourceId)
-    }
-}
-
-
 
 //the same for nested DSL components
 inline fun AnkoContext<*>.dip(value: Int): Int = ctx.dip(value)
@@ -81,10 +57,6 @@ inline fun AnkoContext<*>.sp(value: Float): Int = ctx.sp(value)
 inline fun AnkoContext<*>.px2dip(px: Int): Float = ctx.px2dip(px)
 inline fun AnkoContext<*>.px2sp(px: Int): Float = ctx.px2sp(px)
 inline fun AnkoContext<*>.dimen(@DimenRes resource: Int): Int = ctx.dimen(resource)
-inline fun AnkoContext<*>.dimenAttr(@AttrRes attribute: Int): Int = ctx.dimenAttr(attribute)
-@ColorInt
-inline fun AnkoContext<*>.colorAttr(@AttrRes attribute: Int): Int = ctx.colorAttr(attribute)
-inline fun AnkoContext<*>.attribute(@AttrRes attribute: Int): TypedValue = ctx.attr(attribute)
 
 //the same for the views
 inline fun View.dip(value: Int): Int = context.dip(value)
@@ -94,10 +66,6 @@ inline fun View.sp(value: Float): Int = context.sp(value)
 inline fun View.px2dip(px: Int): Float = context.px2dip(px)
 inline fun View.px2sp(px: Int): Float = context.px2sp(px)
 inline fun View.dimen(@DimenRes resource: Int): Int = context.dimen(resource)
-inline fun View.dimenAttr(@AttrRes attribute: Int): Int = context.dimenAttr(attribute)
-@ColorInt
-inline fun View.colorAttr(@AttrRes attribute: Int): Int = context.colorAttr(attribute)
-inline fun View.attr(@AttrRes attribute: Int): TypedValue = context.attr(attribute)
 
 //the same for Fragments
 inline fun Fragment.dip(value: Int): Int = activity.dip(value)
@@ -107,7 +75,3 @@ inline fun Fragment.sp(value: Float): Int = activity.sp(value)
 inline fun Fragment.px2dip(px: Int): Float = activity.px2dip(px)
 inline fun Fragment.px2sp(px: Int): Float = activity.px2sp(px)
 inline fun Fragment.dimen(@DimenRes resource: Int): Int = activity.dimen(resource)
-inline fun Fragment.dimenAttr(@AttrRes attribute: Int): Int = activity.dimenAttr(attribute)
-@ColorInt
-inline fun Fragment.colorAttr(@AttrRes attribute: Int): Int = activity.colorAttr(attribute)
-inline fun Fragment.attr(@AttrRes attribute: Int): TypedValue = activity.attr(attribute)
